+++ modules/__tests__/toNotEqual-test.js
@@ -1,6 +1,35 @@
 import expect from '../index'
 
 describe('toNotEqual', () => {
+  it('works', () => {
+    expect('actual').toNotEqual('expected')
+  })
+
+  it('works with objects that do not have the same keys', () => {
+    const a = { a: 'a', b: 'b', c: 'c' }
+    const b = { a: 'a', b: 'b', d: 'c' }
+    expect(a).toNotEqual(b)
+  })
+
+  it('works with objects that have different prototypes and different keys', () => {
+    class A {
+      constructor(prop) {
+        this.a = prop
+      }
+    }
+
+    class B {
+      constructor(prop) {
+        this.b = prop
+      }
+    }
+
+    const a = new A('hi')
+    const b = new B('hi')
+
+    expect(a).toNotEqual(b)
+  })
+
   it('works with arrays of objects', () => {
     const a = [
       {